Controller Order & Feedback

+++ SU24_PRN212_SE1717_Group3/Controllers/Dashboard/ProductManagerController.cs
@@ -2,16 +2,22 @@
 using Microsoft.EntityFrameworkCore;
 using DataAccessLayer.DAO;
 using DataAccessLayer.Models;
+using Utilites;
 
 namespace SU24_PRN212_SE1717_Group3.Controllers.Dashboard
 {
-	public class ProductManagerController(ProductDAO productDAO) : Controller
+	public class ProductManagerController(AccountDAO accountDAO, ProductDAO productDAO) : Controller
 	{
-
 		[HttpGet]
 		public async Task<IActionResult> Index(int CurrentPage)
 		{
-			int endpage = (int)Math.Ceiling(productDAO.GetCountProduct() / 6.0);
+			var account = await accountDAO.GetAccountById(HttpContext.Session.GetInt32("accountId"));
+			if (account == null || account.Role.Name != "Shop")
+			{
+				return RedirectToAction("Index", "Home");
+			}
+
+            int endpage = (int)Math.Ceiling(productDAO.GetCountProduct() / 6.0);
 			CurrentPage = CurrentPage < 1 ? 1 : CurrentPage > endpage ? endpage : CurrentPage;
 			var listProduct = await productDAO.GetAllProduct();
 			listProduct = listProduct.Skip((CurrentPage - 1) * 6).Take(6).ToList();
@@ -24,42 +30,73 @@ namespace SU24_PRN212_SE1717_Group3.Controllers.Dashboard
 		[HttpGet]
 		public async Task<IActionResult> Add()
 		{
-			TempData["Category"] = await productDAO.GetCategory();
-			TempData["Shop"] = await productDAO.GetShop();
+            var account = await accountDAO.GetAccountById(HttpContext.Session.GetInt32("accountId"));
+            if (account == null || account.Role.Name != "Shop")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            ViewData["Category"] = await productDAO.GetAllCategory();
 			return View();
 		}
 
 		[HttpPost]
-		public async Task<IActionResult> Add(Product product, int quan, int categoryId, int shopId, IFormFile img)
+		public async Task<IActionResult> Add(Product product, int quantity, int categoryId, int shopId, IFormFile img)
 		{
+            var account = await accountDAO.GetAccountById(HttpContext.Session.GetInt32("accountId"));
+            if (account == null || account.Role.Name != "Shop")
+            {
+                return RedirectToAction("Index", "Home");
+            }
 
-			if (product != null)
-			{
-				await productDAO.AddProduct(product, quan, categoryId, shopId, img);
-			}
+            product.Image = ImgUtil.Compress(ImgUtil.ConvertIFromFileToByte(img));
+			var Category = await productDAO.GetCategoryById(categoryId);
+			await productDAO.AddProduct(product, Category, account.Shop, quantity);
 			return RedirectToAction("Index");
 		}
 
 		[HttpGet]
 		public async Task<IActionResult> Update(int Id)
 		{
-			var product = await productDAO.GetProductById(Id);
-			TempData["Category"] = await productDAO.GetCategory();
-			TempData["Shop"] = await productDAO.GetShop();
-			return View(product);
+            var account = await accountDAO.GetAccountById(HttpContext.Session.GetInt32("accountId"));
+            if (account == null || account.Role.Name != "Shop")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            var product = await productDAO.GetProductById(Id);
+			ViewData["Category"] = await productDAO.GetAllCategory();
+			ViewData["ShopName"] = account.Shop.Name;
+
+            return View(product);
 		}
 
 		[HttpPost]
-		public async Task<IActionResult> Update(Product pro, int quan, string name, IFormFile img)
+		public async Task<IActionResult> Update(Product product, int quantity, int categoryId, IFormFile img)
 		{
-			await productDAO.UpdateProduct(pro, quan, name, img);
+            var account = await accountDAO.GetAccountById(HttpContext.Session.GetInt32("accountId"));
+            if (account == null || account.Role.Name != "Shop")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            product.Image = ImgUtil.Compress(ImgUtil.ConvertIFromFileToByte(img));
+
+			var Category = await productDAO.GetCategoryById(categoryId);
+			await productDAO.UpdateProduct(product, Category, quantity);
 			return RedirectToAction("Index");
 		}
 
 		[HttpGet]
 		public async Task<IActionResult> Delete(int Id)
 		{
-			var pro = await productDAO.GetProductById(Id);
+            var account = await accountDAO.GetAccountById(HttpContext.Session.GetInt32("accountId"));
+            if (account == null || account.Role.Name != "Shop")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            var pro = await productDAO.GetProductById(Id);
 			await productDAO.Delete(pro);
 			return RedirectToAction("Index");
 		}
@@ -67,7 +104,13 @@ namespace SU24_PRN212_SE1717_Group3.Controllers.Dashboard
 		[HttpGet]
 		public async Task<IActionResult> Search(string name, int CurrentPage)
 		{
-			int endpage = (int)Math.Ceiling(productDAO.GetCountProductBySearchName(name) / 6.0);
+            var account = await accountDAO.GetAccountById(HttpContext.Session.GetInt32("accountId"));
+            if (account == null || account.Role.Name != "Shop")
+            {
+                return RedirectToAction("Index", "Home");
+            }
+
+            int endpage = (int)Math.Ceiling(productDAO.GetCountProductBySearchName(name) / 6.0);
 			CurrentPage = CurrentPage < 1 ? 1 : CurrentPage > endpage ? endpage : CurrentPage;
 			var listProduct = await productDAO.GetAllProductBySearchName(name);
 
@@ -79,12 +122,5 @@ namespace SU24_PRN212_SE1717_Group3.Controllers.Dashboard
 			ViewData["CurrentPage"] = CurrentPage;
 			return View("Index", listProduct);
 		}
-
-
-
-
-
-
-
 	}
 }
